requirements.txt
certifi==2017.11.5
chardet==3.0.4
Django==2.0.1
django-materialize==1.0.1a2
idna==2.6
nose==1.3.7
numpy==1.14.0
oauthlib==2.0.6
pandas==0.22.0
python-dateutil==2.6.1
python-dwapi==0.3.dev0
pytz==2017.3
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
urllib3==1.22
